+++ src/proxy/AST/ASTFormater.js
@@ -3,11 +3,11 @@
  * @author zdying
  */
 
-var commandFuncs = require('./commands');
-var merge = require('../helpers/merge');
-var type = require('../helpers/type');
+var commandFuncs = require('./../commands/index');
+var merge = require('../../helpers/merge');
+var type = require('../../helpers/type');
 
-var replaceVar = require('./replaceVar');
+var replaceVar = require('./../tools/replaceVar');
 
 var scopeCmds = {
     domain : ['set'],
@@ -22,6 +22,7 @@ module.exports = function formatAST(ASTTree){
     var res = {
         commands: [],
         props: {},
+        domains: {},
         __id__: 'global'
     };
 
@@ -43,24 +44,55 @@ module.exports = function formatAST(ASTTree){
     // replaceProps(res.props, res);
 
     // step 2: 解析基本规则(比如: `example.com => other.com`)
+    parseBaseRule(baseRules, res);
+
+    parseDomain(domains, res);
+
+    // console.log(JSON.stringify(res, null, 4));
+
+    return res
+};
+
+/**
+ * 将baseRule解析成标准的domain／location对象
+ * @param baseRules
+ * @param res
+ */
+function parseBaseRule(baseRules, res){
     baseRules.forEach(function(rule){
         var arr = rule.split(/\s*=>\s*/);
         var source = arr[0];
         var target = arr[1];
-        var globalProps = res.props;
+        var url = require('url');
 
         // step 3: 替换基本规则中的变量
         source = replaceVar(source, res);
         target = replaceVar(target, res);
 
-        res[source] = {
-            source: source,
-            props: {
-                proxy: target
-            }
-        };
-    });
+        if(!/^[\w\d]+:\/\//.test(source)){
+            source = 'http://' + source;
+        }
 
+        var urlObj = url.parse(source);
+        var hostname = urlObj.hostname;
+
+        if(hostname){
+            res.domains[hostname] = {
+                domain: hostname,
+                location: [
+                    {
+                        path: urlObj.path || '/',
+                        props: {
+                            proxy: target
+                        }
+                    }
+                ]
+            }
+        }
+    });
+}
+
+function parseDomain(domains, res){
     // step 4: 解析Domain(比如: `example.com = { ... }`)
     domains.forEach(function(domain){
         var _domain = domain.domain;
@@ -72,6 +104,7 @@ module.exports = function formatAST(ASTTree){
         domain.toJSON = toJSON;
 
         domain.parent = res;
+        domain.parentID = res.__id__;
 
         // step 5: 执行domain中的命令(比如: set $domain example.com)
         // 这里必须先执行命令, 然后在替换值
@@ -87,7 +120,6 @@ module.exports = function formatAST(ASTTree){
         replaceVar(domain.props, domain);
 
         // funcs里面的变量属于domain, 用domain的变量和上一层变量替换
-
         funcs.forEach(function(fun){
             var params = fun.params;
             var name = fun.name;
@@ -102,71 +134,85 @@ module.exports = function formatAST(ASTTree){
 
         // step 6: 如果没有location, 直接返回domain对象
         if(!Array.isArray(location) || location.length === 0){
-            res[_domain] = {
+            res.domains[_domain] = {
                 source: _domain,
                 commands: funcs,
                 props: domain.props,
                 parent: res,
+                location: [],
                 toJSON: toJSON
             };
             return
+        }else{
+            res.domains[_domain] = domain;
         }
 
-        // step 7: 合并location
-        location.forEach(function(loc){
-            var url = _domain + loc.location;
-            var proxy;
-            var props;
-            var funcs = loc.commands || []; //parseCommand(loc.commands || []);
+        // res[_domain] = {
+        //     _domain: _domain
+        // };
 
-            loc.__id__ = '_location_' + loc.location;
-
-            loc.parent = domain;
-            loc.toJSON = toJSON;
-
-            // step 8: 执行location命令(比如: `set $domain example.com`)
-            execCommand(funcs, loc, 'location');
-
-            loc.props = replaceVar(loc.props, loc);
-
-            // step 9: 替换location变量, 作用域为domain和上层(res)
-            url = replaceVar(url, loc);
-            proxy = replaceVar(loc.props.proxy, loc);
-
-            funcs.forEach(function(fun){
-                var params = fun.params;
-                var name = fun.name;
-
-                if(name === 'set'){
-                    // 如果是 set 命令, 不替换第一个参数
-                    fun.params = [params[0]].concat(replaceVar(fun.params.slice(1), loc))
-                }else{
-                    fun.params = replaceVar(fun.params, loc)
-                }
-
-                // console.log('替换location的function参数:', name, fun.params);
-            });
-
-            props = merge({}, loc.props, {
-                proxy: proxy
-            });
-
-            // 替换正则表达式
-            url = url.replace(/(.*?)~\/(.*)/, '~ /$1$2');
-
-            res[url] = {
-                source: url,
-                commands: funcs,
-                props: props,
-                // location: loc,
-                parent: domain,
-                toJSON: toJSON
-            }
-        })
+        domain.location = [];
+        parseLocation(domain, location, res);
     });
+}
 
-    return res
-};
+function parseLocation(domain, location, res){
+    // step 7: 合并location
+    location.forEach(function(loc){
+        var location = loc.location;
+        var url = domain.domain + location;
+        var proxy;
+        var props;
+        var funcs = loc.commands || []; //parseCommand(loc.commands || []);
+
+        loc.__id__ = '_location_' + location;
+        loc.parent = domain;
+        loc.toJSON = toJSON;
+
+        // step 8: 执行location命令(比如: `set $domain example.com`)
+        execCommand(funcs, loc, 'location');
+
+        loc.props = replaceVar(loc.props, loc);
+        location = replaceVar(location, loc);
+
+        // step 9: 替换location变量, 作用域为domain和上层(res)
+        url = replaceVar(url, loc);
+        proxy = replaceVar(loc.props.proxy, loc);
+
+        funcs.forEach(function(fun){
+            var params = fun.params;
+            var name = fun.name;
+
+            if(name === 'set'){
+                // 如果是 set 命令, 不替换第一个参数
+                fun.params = [params[0]].concat(replaceVar(fun.params.slice(1), loc))
+            }else{
+                fun.params = replaceVar(fun.params, loc)
+            }
+
+            // console.log('替换location的function参数:', name, fun.params);
+        });
+
+        props = merge({}, loc.props, {
+            proxy: proxy
+        });
+
+        // 替换正则表达式
+        url = url.replace(/(.*?)~\/(.*)/, '~ /$1$2');
+
+        domain.location.push({
+            path: location,
+            originPath: loc.location,
+            source: url,
+            commands: funcs,
+            props: props,
+            // location: loc,
+            parent: domain,
+            parentID: domain.__id__,
+            toJSON: toJSON
+        });
+    })
+}
 
 function execCommand(funcs, context, scope){
     if(!Array.isArray(scopeCmds[scope])){
@@ -247,7 +293,7 @@ function toJSON(){
 // console.log(':::formated tree:::');
 // console.log(JSON.stringify(formatedTree, function(key, value){
 //     if(key === 'parent'){
-//         return '__parent__:' + value.__id__;
+//         return undefined;
 //     }else{
 //         return value
 //     }