+++ src/proxy/commands/getCommands.js
@@ -5,17 +5,17 @@
 
 'use strict';
 
-var types = require('./commands/type');
+var types = require('./type');
 
 /**
- * 获取source对象对应的指令以及祖先元素的指令，并根据type过滤
- * @param {Object} source 源对象
+ * 获取rewrite rule对象对应的指令以及祖先元素的指令，并根据type过滤
+ * @param {Object} rewrite_rule 源对象
  * @param {String} [type] 过滤的类型
  * @returns {Array}
  */
-module.exports = function getCommonds(source, type){
+module.exports = function getCommonds(rewrite_rule, type){
     var tmp = [];
-    var curr = source;
+    var curr = rewrite_rule;
     var currCMDs = [];
     var typedCmds = type && types[type];