+++ src/proxy/index.js
@@ -3,4 +3,4 @@
  * @author zdying
  */
 
-module.exports = require('./Server');
+module.exports = require('./server/Server');