+++ test/proxy.rewrite.test.js
@@ -9,14 +9,15 @@ var assert = require('assert');
 var path = require('path');
 var fs = require('fs');
 
-var parseHosts = require('../src/proxy/parseHosts');
+var parseHosts = require('../src/proxy/tools/parseHosts');
 
 describe('proxy rewrite',function(){
-    var formatAST = require('../src/proxy/formatAST');
-    var AST = require('../src/proxy/AST');
+    var formatAST = require('../src/proxy/AST/ASTFormater');
+    var AST = require('../src/proxy/AST/AST');
 
     it('正确解析AST', function(){
         var sourceCode = fs.readFileSync(__dirname + '/proxy/rewrite.example');
+        // var sourceCode = fs.readFileSync(__dirname + '/proxy/rewrite1.example');
         var rules = AST(sourceCode);
 
         var target = require('./proxy/ASTTree.result');
@@ -25,7 +26,7 @@ describe('proxy rewrite',function(){
     });
 
     it('正确解析格式化AST Tree', function(){
-        var sourceCode = fs.readFileSync(__dirname + '/proxy/rewrite.example');
+        var sourceCode = fs.readFileSync(__dirname + '/proxy/rewrite1.example');
         var rules = AST(sourceCode);
         var formatedTree = formatAST(rules);